+++ Tests/Unit/Fixtures/LoadableClass.php
@@ -1,0 +1,11 @@
+<?php
+namespace T3v\T3vPage\Tests\Unit\Fixtures;
+
+/**
+ * Loadable Class
+ */
+class LoadableClass {
+  public function returnsTrue() {
+    return true;
+  }
+}